gundua_new_backend/requirements.txt
alabaster==0.7.12
ansible-base==2.10.8
apache-libcloud==3.2.0
#apturl==0.5.2
argcomplete==1.10.3
asgiref==3.5.2
asttokens==2.1.0
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.8.2
bert-extractive-summarizer==0.10.1
bitarray==2.6.0
blinker==1.4
blis==0.7.9
bootstrap-py==1.0.2
#Brlapi==0.8.3
bson==0.5.10
cassandra-driver==3.25.0
catalogue==2.0.8
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
#chrome-gnome-shell==0.0.0
click==8.1.3
colorama==0.4.4
command-not-found==0.3
compressed-rtf==1.0.6
confection==0.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
#cupshelpers==1.0
cymem==2.0.7
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
#defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
dj-rest-auth==2.2.6
Django==3.1.14
django-allauth==0.52.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.19
docx2txt==0.8
drf-yasg==1.21.4
#duplicity==0.8.21
ebcdic==1.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
extract-msg==0.28.7
fasteners==0.14.1
filelock==3.8.0
future==0.18.2
fuzzywuzzy==0.18.0
geomet==0.2.1.post1
gitdb==4.0.10
GitPython==3.1.30
httpie==3.2.1
httplib2==0.20.2
huggingface-hub==0.10.1
idna==3.4
imagesize==1.4.1
IMAPClient==2.1.0
importlib-metadata==4.6.4
inflection==0.5.1
ipykernel==6.17.0
ipython==8.6.0
itypes==1.2.0
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonschema==3.2.0
jupyter_client==7.4.4
jupyter_core==4.11.2
keyring==23.5.0
langcodes==3.3.0
#language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
#louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
murmurhash==1.0.9
mysql-connector-python==8.0.31
nest-asyncio==1.5.6
netaddr==0.8.0
netifaces==0.11.0
nltk==3.7
node==1.0
npm==0.1.1
ntlm-auth==1.4.0
numpy==1.23.4
oauthlib==3.2.0
odict==1.9.0
olefile==0.46
openpyxl==3.0.10
optional-django==0.1.0
packaging==21.3
pandas==1.5.0
paramiko==2.9.3
parso==0.8.3
pathy==0.6.2
paytmchecksum==1.7.0
pdfminer.six==20191110
pdfreader==0.1.12
pexpect==4.8.0
pguard==0.4.0
pickleshare==0.7.5
Pillow==9.2.0
pipenv==2022.11.11
platformdirs==2.5.1
plumber==1.7
pre==1.3.0
preshed==3.0.8
prompt-toolkit==3.0.31
protobuf==3.12.4
psutil==5.9.3
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycryptodome==3.15.0
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==1.10.2
pydub==0.25.1
pygame==2.1.2
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pykerberos==1.1.14
pymacaroons==0.13.0
pymongo==4.3.2
PyNaCl==1.5.0
pyparsing==3.0.9
PyPDF2==2.11.1
pyRFC3339==1.1
pyrsistent==0.16.0
PySocks==1.7.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-docx==0.8.11
python-dotenv==0.21.0
python-Levenshtein==0.12.2
python-pptx==0.6.21
python3-openid==3.2.0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
pywinrm==0.3.0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
regex==2022.9.13
reportlab==3.6.8
requests==2.28.1
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
resolvelib==0.8.1
rich==13.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.1.2
scipy==1.9.3
SecretStorage==3.3.1
selinux==3.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
simplejson==3.17.6
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.4.2
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SpeechRecognition==3.8.1
Sphinx==6.1.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib.applehelp==1.0.3
sqlparse==0.4.3
srsly==2.4.5
stack-data==0.6.0
systemd-python==234
tabulate==0.9.0
textract==1.6.5
texttable==1.6.7
thinc==8.1.5
threadpoolctl==3.1.0
tokenizers==0.13.1
torch==1.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.23.1
typer==0.4.2
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uplink==0.9.7
uritemplate==4.1.1
urllib3==1.26.12
usb-creator==0.3.7
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
wadllib==1.3.6
wasabi==0.10.1
wcwidth==0.2.5
websocket-client==0.59.0
whitenoise==5.2.0
wifite==2.6.0
xdg==5
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==3.0.3
xmltodict==0.12.0
zipp==1.0.0
zope.component==5.0.1
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.4
zope.proxy==4.5.0
